renamed _types to service_types
